refactor: :art: agruping domain common entities
agruping domain common entities under "common" folder
