+++ common/infrastructure/utils/const.go
